+++ lib/dealog_backoffice/accounts/projections/account.ex
@@ -8,10 +8,15 @@ defmodule DealogBackoffice.Accounts.Projections.Account do
   schema "accounts" do
     field :first_name, :string
     field :last_name, :string
-    field :administrative_area, :string
+    field :administrative_area_id, :string
     field :organization, :string
     field :position, :string
+
     belongs_to :user, DealogBackoffice.Accounts.User
+
+    has_one :administrative_area, DealogBackoffice.AdministrativeAreas.AdministrativeArea,
+      foreign_key: :ars,
+      references: :administrative_area_id
 
     timestamps()
   end