+++ priv/repo/migrations/20201123101328_add_administrative_area_association_to_account_projection.exs
@@ -1,0 +1,7 @@
+defmodule DealogBackoffice.Repo.Migrations.AddAdministrativeAreaAssociationToAccountProjection do
+  use Ecto.Migration
+
+  def change do
+    rename table(:accounts), :administrative_area, to: :administrative_area_id
+  end
+end